Add click counting functionality
Added click counter functionality that tracks the amount of clicks on the window in the form of the 'data-click-count' attribute on the body element, as well as the '--global-click-count' custom property on the body element. 

This functionality also keeps track of "local" clicks, saving the amount of clicks done on en element through its own 'data-click-count' attribute and '--click-count' custom css property.